phase-2/backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
sqlmodel
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
psycopg2-binary
alembic
asyncpg
python-dotenv
pytest
httpx
jose
cryptography
slowapi
sqlmodel